Add keyboard and pointer device support to Window_screen module; refactor device initialization
